refactored all the things
