+++ packages/rocket-auth-aws-infrastructure/src/lambdas/challenge-create.ts
