+++ packages/rocket-auth-aws-infrastructure/src/lambdas/challenge-define.ts
