+++ packages/rocket-auth-aws-infrastructure/src/lambdas/challenge-verify.ts
