+++ packages/rocket-auth-aws-infrastructure/src/lambdas/confirm-forgot-password.ts
@@ -1,18 +1,21 @@
 import { CognitoIdentityServiceProvider } from 'aws-sdk'
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
-import { errorResponse, response } from '../response'
+import { errorResponse, response } from './response'
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   try {
     const params = JSON.parse(event.body!)
     const cognitoService = new CognitoIdentityServiceProvider()
     await cognitoService
-      .globalSignOut({
-        AccessToken: params.accessToken,
+      .confirmForgotPassword({
+        ClientId: process.env.userPoolClientId!,
+        Username: params.username,
+        ConfirmationCode: params.code,
+        Password: params.password,
       })
       .promise()
     return response(200, {
-      message: 'Signed out',
+      message: 'Your password has been changed',
     })
   } catch (e) {
     return errorResponse(e)
