+++ packages/rocket-auth-aws-infrastructure/src/lambdas/sign-out.ts
@@ -1,19 +1,18 @@
 import { CognitoIdentityServiceProvider } from 'aws-sdk'
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
-import { errorResponse, response } from '../response'
+import { errorResponse, response } from './response'
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   try {
     const params = JSON.parse(event.body!)
     const cognitoService = new CognitoIdentityServiceProvider()
     await cognitoService
-      .forgotPassword({
-        ClientId: process.env.userPoolClientId!,
-        Username: params.username,
+      .globalSignOut({
+        AccessToken: params.accessToken,
       })
       .promise()
     return response(200, {
-      message: "We have sen't you a confirmation code to change your password.",
+      message: 'Signed out',
     })
   } catch (e) {
     return errorResponse(e)
