+++ packages/rocket-auth-aws-infrastructure/src/lambdas/sign-up-confirm.ts
@@ -1,21 +1,20 @@
 import { CognitoIdentityServiceProvider } from 'aws-sdk'
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
-import { errorResponse, response } from '../response'
+import { errorResponse, response } from './response'
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   try {
     const params = JSON.parse(event.body!)
     const cognitoService = new CognitoIdentityServiceProvider()
     await cognitoService
-      .confirmForgotPassword({
+      .confirmSignUp({
         ClientId: process.env.userPoolClientId!,
         Username: params.username,
         ConfirmationCode: params.code,
-        Password: params.password,
       })
       .promise()
     return response(200, {
-      message: 'Your password has been changed',
+      message: `Username: ${params.username} confirmed.`,
     })
   } catch (e) {
     return errorResponse(e)
